milestone3/requirements.txt
comet_ml==3.35.2
Flask==2.2.2
Werkzeug==2.2.2
joblib==1.2.0
matplotlib==3.7.2
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
scikit_learn==1.2.2
streamlit==1.29.0
gunicorn==21.2.0
seaborn==0.12.2